+++ Assignment - Operazioni Matematiche/src/model/Sottrazione.java
@@ -7,7 +7,7 @@ public class Sottrazione extends Operation{
 	}
 	
 	@Override
-	protected int doOperation() {
+	protected double doOperation() {
 		return this.a-this.b;
 	}
